lambda/aws-rag-appsync-stepfn-opensearch/s3_file_transformer/src/requirements.txt
aws-lambda-powertools
aws-xray-sdk
fastjsonschema
typing-extensions
boto3>=1.34.29
requests
langchain==0.1.4
pypdf2
Pillow
langchain-community==0.0.16
